requirements.txt
# Core dependencies
SQLAlchemy>=2.0.0
requests>=2.32.0
feedparser>=6.0.0
google-api-python-client>=2.187.0
beautifulsoup4>=4.14.0
python-dotenv>=1.2.0
pydantic-settings>=2.12.0
click>=8.3.0
python-dateutil>=2.9.0

# Development dependencies (optional)
alembic>=1.17.0  # for database migrations
pytest>=8.0.0    # for testing

# UI enhancement
colorama>=0.4.6
